feat(PX-1176): indices fix

+++ packages/jira/src/indices/indices.ts
@@ -226,8 +226,8 @@ const indices = [
             timeTracker: {
               type: 'object',
               properties: {
-                estimate: { type: 'integer' },
-                actual: { type: 'integer' },
+                estimate: { type: 'long' },
+                actual: { type: 'long' },
               },
             },
           },
